+++ packages/calling-component-bindings/src/baseSelectors.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { toFlatCommunicationIdentifier } from 'acs-ui-common';
+import { toFlatCommunicationIdentifier } from '@internal/acs-ui-common';
 import { CallState, CallClientState, DeviceManagerState, IncomingCallState } from 'calling-stateful-client';
 
 /**