+++ packages/chat-component-bindings/src/chatParticipantListSelector.ts
@@ -3,7 +3,7 @@
 
 import { getUserId, getDisplayName, getParticipants } from './baseSelectors';
 import * as reselect from 'reselect';
-import { toFlatCommunicationIdentifier } from 'acs-ui-common';
+import { toFlatCommunicationIdentifier } from '@internal/acs-ui-common';
 import { ChatParticipant } from '@azure/communication-chat';
 import { CommunicationParticipant } from 'react-components';