+++ packages/chat-component-bindings/src/chatThreadSelector.ts
@@ -8,9 +8,9 @@ import {
   getUserId,
   sanitizedMessageContentType
 } from './baseSelectors';
-import { toFlatCommunicationIdentifier } from 'acs-ui-common';
+import { toFlatCommunicationIdentifier } from '@internal/acs-ui-common';
 import { ChatMessageWithStatus } from 'chat-stateful-client';
-import { memoizeFnAll } from 'acs-ui-common';
+import { memoizeFnAll } from '@internal/acs-ui-common';
 import { ChatMessage, MessageAttachedStatus } from 'react-components';
 import { createSelector } from 'reselect';
 import { compareMessages } from './utils/compareMessages';